requirements.txt
Django
dj_static
djangorestframework
gunicorn
pandas